+++ src/reddit_topics_aggregator/package_metadata.py
@@ -1,2 +1,2 @@
 __title__ = "reddit-topics-aggregator"
-__version__ = "0.1.0"
+__version__ = "0.2.0"