+++ CHANGELOG.md
@@ -5,18 +5,33 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## TODO
+
+- add support for json output
+- add support for markdown output
+- add count of articles if duplicate
+- remove pinned topics
+
 ## [Unreleased]
 
 ### Added
-
-- environment variable defaults are now documented and used
-- command `topics` - CLI sub-command to retrieve and display topic submissions for the specified subreddits
 
 ### Fixed
 
 ### Changed
 
 ### Removed
+
+## [0.2.0]
+
+### Added
+
+- environment variable defaults are now documented and used
+- command `topics` - CLI sub-command to retrieve and display topic submissions for the specified Subreddits
+
+### Fixed
+
+- coverage in Github was failing with "import file mismatch"
 
 ## [0.1.0]
 
@@ -33,5 +48,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial Release
 
-[unreleased]: https://github.com/lieutdan13/reddit-topics-aggregator/compare/v0.0.1...HEAD
+[unreleased]: https://github.com/lieutdan13/reddit-topics-aggregator/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/lieutdan13/reddit-topics-aggregator/releases/tag/v0.2.0
+[0.1.0]: https://github.com/lieutdan13/reddit-topics-aggregator/releases/tag/v0.1.0
 [0.0.1]: https://github.com/lieutdan13/reddit-topics-aggregator/releases/tag/v0.0.1